+++ test/WebJobs.Script.Tests.Integration/WebHostEndToEnd/SamplesEndToEndTests_Node.cs
@@ -361,11 +361,6 @@ namespace Microsoft.Azure.WebJobs.Script.Tests.EndToEnd
                 });
             }
 
-            public override async Task DisposeAsync()
-            {
-                await Host.StopAndDisposeWebHostAsync();
-                await base.DisposeAsync();
-            }
         }
     }
 }